project/deeplearning/segment/BiseNetV2/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.2
opencv-python>=4.4.0
pillow>=8.0.0
scipy>=1.5.2
tqdm>=4.50.0
tensorboardX>=2.1
matplotlib>=3.3.2
albumentations>=0.5.2